+++ lib/prop_flex.ml
@@ -120,7 +120,7 @@ let normalize_flex (value : flex) : flex =
 let rec pp_order : order Pp.t =
  fun ctx -> function
   | Int i -> Pp.int ctx i
-  | Calc c -> pp_calc pp_order ctx c
+  | Calc c -> pp_calc ~unwrap_num:false pp_order ctx c
   | Inherit -> Pp.string ctx "inherit"
   | Initial -> Pp.string ctx "initial"
   | Unset -> Pp.string ctx "unset"